usb: dwc3: gadget: disable the uneeded PHY
on connection done IRQ we already know the
negotiated speed and that won't change at
least until we disconnect and connect the
cable. So we can Suspend the unnecessary
PHY on that handler and save some power.

on reset IRQ we reenable both PHYs so
that we can negotiate our speed.

Signed-off-by: Felipe Balbi <balbi@ti.com>