usb: dwc3: gadget: fix maxpacket size for SS
wMaxPacketSize is directly re-used as bMaxPacketSize0 for the
USB_DT_DEVICE request. While the result is okay for LS/FS/HS the
parameter changed in USB3.0: it is 2^X instead X.
While here also change the initial parameter from -EINVAL to 0. It
should live short since it is overwritten during probe where we start in
SS.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Felipe Balbi <balbi@ti.com>