usb: dwc3: gadget: fix starting one EP more than once
After a transfer complete we could start a new transfer
if something was enqueued and we completed the last one.

If the we received a TRANSFER_NOT_READY then we could start
another transfer for the same EP. This was working often but
sometimes it lead to errors. It is wrong anyway.

The logic was changed recently to start transfers once we
receive TRANSFER_NOT_READY because if we start them right
away we could overrun windows with requests. For that we
now use DWC3_EP_BUSY flag to signalize that the EP is busy
i.e. performing one or more transfers. In that situation we
don't start any transfers; instead we wait until the transfer
completes.

We still start new transfers after we completed a batch of
those even if we did not receive a TRASNFER_NO_READY yet.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Felipe Balbi <balbi@ti.com>