usb: dwc3: gadget: fix multi TRB handling
So I assumed we get a XferComplete event for all transfers but only the
interrupt is not triggered. Test showed that this is not the case. What
happens is that we only receive on event and one interrupt for all
tranfers. Therefore we have to cleanup our complete list with transfers
which have been scheduled.
While here we also interprete the BUS ERROR bit. If it is set I assume
that all tranfers failed.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Felipe Balbi <balbi@ti.com>