usb: dwc3: gadget: don't set IOC bit
I've done some testing and come to the conclusion that we don't need to
set this bit at all for non-ISOC transfers.
If we set the LST bit than the IOC is ignored and serves no purpose.
If LST is not set then IOC will generate an XferInProgress event
including an Interrupt. So this seems to be handy for IOC where we don't
use the LST bit in order to use a ring which is "always" scheduled.
Therefore remove setting this bit.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Felipe Balbi <balbi@ti.com>