usb: dwc3: gadget: don't queue transfers if we got interrupt on a short packet
If we have 10 TRBs in queued for transfers (OUT) then only the last one
has the LST bit set meaning only after the last one completes, we
receive a XferComplete complete event.
However, we also receive a XferComplete on a short packet. If we start
the queue again and enqueue requests that may have been enqueued in the
meantime (like in complete callback) then we would put new TRBs on the
queue with LST bit set. Since we try to avoid, lets skip that until we
run out of TRBs.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Felipe Balbi <balbi@ti.com>