usb: dwc3: gadget: handle more cases in the clean up routine
Right now we consider all requests as done once we receive XferComplete.
We only complete one request if the "short" bit in the event is set.
Now, lets extend this:

- stop once the current TRB has the LST bit set. In theorhy we have
  multiple transfers running. Every LST bit starts one XferComplete
  event.

- Consider the SHORT bit ony if the transfer is less that we asked for.
  We could have two TRB queued where the first one was complete and the
  second was short. The first creates no event because we did not set
  LST or IOC or anything. The second TRB creates the event and that is
  the position where we break.

In both cases the queue is halted and we should start it again if we
have requests pending on one of the lists.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Felipe Balbi <balbi@ti.com>