usb: dwc3: gadget: make the pre-mature dequeue of requests possible
Active requests become a valid TRB assigned and the START_TRANSFER
command is issued. If the gadget decides to remove the requests we
return -EBUSY instead of removing the requests. The gadget _could_
respect this. However if the gadget is removed (rmmod for instance) we
have to clean up everything. In that case -EBUSY isn't helping.
This ensures that the armed TRB is removed. We probably should add some
kind of synchronization so ep_disable() waits until the completion of
the ep command which calls complete or else we might call ->complete()
after the driver is gone.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Felipe Balbi <balbi@ti.com>