ODROID-X: lcd: Remove vertical lines of LP101WH1
This patch to remove the vertical lines at right side of lcd panel. The lines
are drawen if the frame width is not align the sum of double-word (8 bytes)
boundary. The width of LP101WH1 is 1366 and it's not aligned with double-word
boundary, therefore we only use 1360 pixels, another 6 pixels are ignored.

Change-Id: I5b780c93bbcd2526b7d7ecfd824377610c1b0f6f
Signed-off-by: Dongjin Kim <dongjin.kim@agreeyamobility.net>