loop: limit the number of requests in the bio list
Currently there is no limit on the number of requests in the loop bio
list.  This can lead into some nasty situations when the caller spawns
tons of bio requests taking huge amount of memory.  This is even more
obvious with discard where blkdev_issue_discard() will submit all bios for
the range and wait for them to finish afterwards.  On really big loop
devices and slow backing file system this can lead to OOM situation as
reported by Dave Chinner.

With this patch we will wait in loop_make_request() if the number of bios
in the loop bio list would exceed 'nr_congestion_on'.  We'll wake up the
process as we process the bios form the list.  Some threshold hysteresis
is in place to avoid high frequency oscillation.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>
Reported-by: Dave Chinner <david@fromorbit.com>
Cc: Neil Brown <neilb@suse.de>
Cc: David Howells <dhowells@redhat.com>
Cc: Ingo Molnar <mingo@elte.hu>
Cc: Peter Zijlstra <a.p.zijlstra@chello.nl>
Cc: Jens Axboe <axboe@kernel.dk>
Acked-by: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>