aoe: improve handling of misbehaving network paths
An AoE target can have multiple network ports used for AoE, and in the aoe
driver, those are tracked by the aoetgt struct.  These changes allow the
aoe driver to handle network paths, or aoetgts, that are not working well,
compared to the others.

Paths that do not get responses despite the retransmission of AoE commands
are marked as "tainted", and non-tainted paths are preferred.

Meanwhile, the aoe driver attempts to "probe" the tainted path in the
background by issuing reads of LBA 0 that are padded out to full (possibly
jumbo-frame) size.  If the probes get responses, then the path is
"redeemed", and its taint is removed.

This mechanism has been shown to be helpful in transparently handling and
recovering from real-world network "brown outs" in ways that the earlier
"shoot the help-needing target in the head" mechanism could not.

Signed-off-by: Ed Cashin <ecashin@coraid.com>
Cc: Jens Axboe <axboe@kernel.dk>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>